devtools/server/tests/chrome/test_memory_allocations_02.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1132764 - Test controlling the maximum allocations log length over the RDP.
-->
<head>
  <meta charset="utf-8">
  <title>Memory monitoring actor test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="memory-helpers.js" type="application/javascript"></script>
<script>
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();

  (async function() {
    const { memory, target } = await startServerAndGetSelectedTabMemory();
    await memory.attach();

    const allocs = [];
    let eventsFired = 0;
    let intervalId = null;
    function onAlloc() {
      eventsFired++;
    }
    function startAllocating() {
      intervalId = setInterval(() => {
        for (let i = 100000; --i;) {
          allocs.push({});
        }
      }, 1);
    }
    function stopAllocating() {
      clearInterval(intervalId);
    }

    memory.on("allocations", onAlloc);

    await memory.startRecordingAllocations({
      drainAllocationsTimeout: 10,
    });

    await waitUntil(() => eventsFired > 5);
    ok(eventsFired > 5,
       "Some allocation events fired without allocating much via auto drain");
    await memory.stopRecordingAllocations();

    // Set a really high auto drain timer so we can test if
    // it fires on GC
    eventsFired = 0;
    const startTime = performance.now();
    const drainTimer = 1000000;
    await memory.startRecordingAllocations({
      drainAllocationsTimeout: drainTimer,
    });

    startAllocating();
    await waitUntil(() => {
      Cu.forceGC();
      return eventsFired > 1;
    });
    stopAllocating();
    ok(performance.now() - drainTimer < startTime,
       "Allocation events fired on GC before timer");
    await memory.stopRecordingAllocations();

    memory.off("allocations", onAlloc);
    await memory.detach();
    destroyServerAndFinish(target);
  })();
};
</script>
</pre>
</body>
</html>
